+++ src/server/routes/api.js
@@ -3,6 +3,6 @@ import Express from 'express';
 const router = Express.Router();
 
 router.use('/4chan', require('./4chan'));
-router.use('/reddit', ()=>log.warn("Reddit request. make provider handler"));
+router.use('/reddit', require('./reddit'));
 
 export default router